It's been forever since I've committed, but I know I just added landscape mode support, and tightly integrated internal framework "QutoeKit", and made a WatchKit target

+++ Quoted/MHAlertBannerView.h
@@ -7,7 +7,7 @@
 //
 
 @import UIKit;
-typedef NS_ENUM(int16_t, MHAlertBannerViewStyle){
+typedef NS_ENUM(NSInteger, MHAlertBannerViewStyle){
     MHAlertBannerViewStyleFacebookPost,
     MHAlertBannerViewStyleTwitterPost,
     MHAlertBannerViewStyleCustom
